refactor: `utils/api/index.ts` exporta todo `utils/api`
