+++ frontend/src/components/navigation/CanchaMenu.tsx
@@ -3,7 +3,7 @@ import { useEstablecimientoByID } from "../../utils/api/establecimientos";
 import { useParams } from "react-router";
 import LoadingSpinner from "../LoadingSpinner/LoadingSpinner";
 import Alerta from "../Alerta/Alerta";
-import { useCanchaByID } from "@/utils/api/canchas";
+import { useCanchaByID } from "@/utils/api";
 
 export default function CanchaMenu() {
   const { idEst, idCancha } = useParams();
