+++ frontend/src/hooks/useCurrentJugador.tsx
@@ -4,7 +4,7 @@ import { readLocalStorage } from "@/utils/localStorage";
 import jwtDecode from "jwt-decode";
 import { JWT } from "@/utils/api";
 import { useToast } from "@chakra-ui/react";
-import { refreshToken } from "@/utils/api/auth";
+import { refreshToken } from "@/utils/api";
 
 interface ICurrentJugadorContext {
   jugador: Jugador;
