+++ frontend/src/pages/ests/[idEst]/canchas/[idCancha]/disps/_FormEliminarDisp.tsx
@@ -1,6 +1,6 @@
 import { ConfirmSubmitButton } from "@/components/forms";
 import { useParams } from "@/router";
-import { useEliminarDisponibilidad } from "@/utils/api/disponibilidades";
+import { useEliminarDisponibilidad } from "@/utils/api";
 import { DeleteIcon } from "@chakra-ui/icons";
 import { useToast } from "@chakra-ui/react";
