+++ frontend/src/pages/jugador/[idJugador]/est/[idEst]/canchas/[idCancha]/index.tsx
@@ -19,12 +19,12 @@ import {
   Td,
   TableContainer,
 } from "@chakra-ui/react";
-import { useCanchaByID } from "@/utils/api/canchas";
+import { useCanchaByID } from "@/utils/api";
 import { useParams } from "@/router";
 
 import React, { useState, useEffect } from "react";
 import { DIAS_ABBR, FALLBACK_IMAGE_SRC } from "@/utils/consts/consts";
-import { useEstablecimientoByID } from "@/utils/api/establecimientos";
+import { useEstablecimientoByID } from "@/utils/api";
 import LoadingSpinner from "@/components/LoadingSpinner/LoadingSpinner";
 import FormReservarDisponibilidad from "./_formReservar";
 import { ordenarDias } from "@/utils/dias";
