+++ frontend/src/pages/jugador/[idJugador]/est/[idEst]/canchas/index.tsx
@@ -1,7 +1,7 @@
 import { useParams } from "react-router";
 import { Box, HStack, Heading } from "@chakra-ui/react";
-import { useEstablecimientoByID } from "@/utils/api/establecimientos";
-import { useCanchasByEstablecimientoID } from "@/utils/api/canchas";
+import { useEstablecimientoByID } from "@/utils/api";
+import { useCanchasByEstablecimientoID } from "@/utils/api";
 import CanchaJugador from "@/components/EstablecimientosJugador/CanchaJugador";
 import { useLocation } from "react-router";
 import { formatearFecha } from "@/utils/dates";
